2025.01.03

+++ lib/ui/screens/transactions/transactions.dart
@@ -1,6 +1,7 @@
 import 'package:az_cash/database/controllers/transaction_controller.dart';
 import 'package:az_cash/database/models/transaction.dart';
 import 'package:az_cash/ui/components/order_widget.dart';
+import 'package:az_cash/ui/constant.dart';
 import 'package:az_cash/ui/helper/widget_helper.dart';
 import 'package:az_cash/ui/screens/transactions/components/delete_order.dart';
 import 'package:flutter/material.dart';
@@ -23,16 +24,19 @@ class TranasctionsScreen extends StatelessWidget {
               itemBuilder: (context, index) {
                 Transaction? order = orders[index];
                 if (order == null) return const SizedBox();
+
                 return GestureDetector(
                     behavior: HitTestBehavior.opaque,
                     onLongPress: () async {
-                      showModalBottomSheet(
-                        context: context,
-                        isScrollControlled: true,
-                        builder: (context) {
-                          return const DeleteOrder();
-                        },
-                      );
+                      if (AppData.user!.role == "Master") {
+                        showModalBottomSheet(
+                          context: context,
+                          isScrollControlled: true,
+                          builder: (context) {
+                            return const DeleteOrder();
+                          },
+                        );
+                      }
                     },
                     child: OrderWidget(order: order));
               },
